Fix for occasionally return no data when socket connection is slow, interrupted, or delayed (MySql bug #69039, Oracle bug #16950212).
This fix retry to read socket package after a TimeoutException, specially when used Ping command, also closes Driver to avoid using socket.
The problem is when Connector/NET tries to get the response (for example after a Ping or ChangeDatabase command) and a socket timeout is thrown, if the next statement is a COM_QUERY, the socket reads the previous response (ok packet) and the query result is taken as no data response.