WI #1132 [DBGeometry type] DBGeometry type used in DbContext is not working.
Renamed MySql.Data.Entity.dll to MySql.Data.Entity.EF5.dll for Entity Framework 5 and .NET Framework 4.5 or above. 
This change was done because there is a conflict between MySql.Data.Entity for .NET Framework 4.0 installed in the GAC and the assembly for .NET Framework 4.5. As both assemblies have the same name, by default the one in the GAC is always loaded and that version has no support for Spatial types. In order to load the assembly for 4.5, the name must be different.