Fix for Bug #27251839 NuGet packages could not be installed from within Visual Studio 2015.
Removed System.Runtime reference in MySql.Data.csproj and MySql.Data.Entity.EF6.csproj files because of a bug with Nuget and VS2015:

https://github.com/NuGet/Home/issues/4639

Removed xml doc files and transformation files for MySql.Data Nuget package.
Transformation files are applied when using packages.config only and they are not applied using PackageReference (in this scenario they are added to project as content files).