Bug #33556024 - Connector/Net doesn't recognize utf8mb3 Added the "utf8mb3" to the supported charsets.
Updated copyright year of README files.

Bug #20266825 - SQLNULLVALUEEXCEPTION THROWN WHEN CALLING MYSQLCONNECTION::GETSCHEMA
When trying to retrieve the value of the "Collation" column of the index table, it was raising an error because of a NULL value. So instead of getting the string directly, it is now getting a generic value so it could handle a NULL value.