Bug #23343947 - .NET BUG WRITE NULLABLE VALUES At the moment when a parameter was created, the default value of the DbType and MySQLDbType was used, hence the difference between them. The MySqlParameter constructor was modified to set the same type (String) for both, DbType and MySQLDbType.
Bug #20056757 - MYSQLPARAMETER.CLONE MISSED ASSIGN VALUE TO PROPERTY SOURCECOLUMNNULLMAPPING
Added the SourceColumnNullMapping property into the constructor used by the Clone method.

Bug #28824591 - MYSQL NET CONNECTOR 8.0.12 SENDS ILLEGAL CONNECT ATTRIBUTES.
Fixed how the length of the connection attributes are calculated.
The length of the string value is not always the same as the length of the bytes when encoded to UTF-8 (C/NET always convert to UTF-8 for serialization across the wire), i.e. chinese characters.

Bug #32050204 - DEFAULT VALUE FOR MYSQLPARAMETER.VALUE CHANGED FROM NULL TO 0
The fix for this bug consisted in reverting the work done for https://mybug.mysql.oraclecorp.com/orabugs/25573071 since it was wrong implemented.
C# has a implicit conversion to Enum when the value is 0, that's something inevitable for us to avoid (see https://stackoverflow.com/questions/2043554/method-overload-resolution-unexpected-behavior).
C/NET offers different ways to create a parameter with value equals to zero, so instead of forcing this we are not going to mess around the natural behavior of C#.

Bug #24495619 - MySqlDateTime type bug in .NET provider
Added missing implementation of the IConvertible interface