Bug #33470147 - Bug #91752 is marked as fixed in v8.0.16, but still present in v8.0.26 When reading a zero time value, it didn't reset the value of the new row hence the exception. If at the moment of reading the value detects it is null, then set value as '00:00:00'; just as a new created type.
Bug #20328929 - EXTRA SELECT@@SQL_MODE QUERIES IN BATCHED INSERT
When performing a batched insert using MySqlDataAdapter, the MySqlCommand.GetCommandTextForBatching() method was querying the "@@sql_mode" for every insert. This extra consult was removed and instead the Property "sql_mode" from the driver is used since this is obtained at the begginig of the connection.