Bug #22913833 - COMMANDS IGNORED AND NO ERROR PRODUCED WHEN PACKET OVER MAX_ALLOWED_PACKET When doing a batch insert and the packet exceeded the 'max_allowed_packet', it just skip the command and passes to the next one. This behavior wasn't the ideal since partial batches shouldn't apply. Instead of skipping the command, an exception is thrown.
Bug #28980953 - MYSQLDATAREADER.GETSTREAM THROWS INDEXOUTOFRANGEEXCEPTION
When trying to execute GetStream method of a MySqlDataReader, the connector was raising an exception since the method was not actually implemented, this was trying to access to the base class method (DbDataReader).
Implemented the GetStream(int i) method in the MySqlDataReader class, which will return a Stream. This is only valid for Binary and GUID type columns.

Bug #32933120 - CAN'T USE TIMESPAN VALUE WITH MICROSECONDS USING PREPARED STATEMENT
At the moment of writing the time value, the calculation of microseconds was wrong implemented