WL#15193 [Support WebAuthn in fido authentication plugin [Classic]]
This patch implemented the support for WebAuthn authentication plugin. A new method from 'libfido2' library has been mapped to check if the FIDO device supports CTAP 2.1 Credential Management.
A new method from 'libfido2' library has been mapped to get the number of statements in an assert. The client now sends back to the server all of the assertions returned by the authenticator device instead of only sending the first one.

Change-Id: Id2865c2403ffd201f6cb958f1ab8c36f5fbfe6e0