WIP Ensure sql statements end up in ActivitySource
There are quite a few ways for an sql statement to wound up being
executed, including but not only:

1. Passing it into the `MySQLCommand` constructor
2. Assigning a `MySQLCommand` instance's `CommandText`
3. Using `CommandType.TableDirect` with a table name
4. Using a stored procedure
5. ...with parameters
6. Opening a connection

And probably more. The existing bridge into the `ActivitySource` answers
some of them, and this commit hopes to bridge some of the gap. Namely:

- Scenario 2 (`CommandText` assignment)
- Scenario 3 (`TableDirect`) shows the actually executing SQL

This commit is currently to be considered in a draft state; it is not
ready to be merged. Inline review comments will be added for
clarifications and questions.